Remove the profiles for LWJGL libs
I wanted to use this lib in a test project, where I wanted to open a GameWindow.
The problem was, that by defining the Java Simple Game Engine dependency inside
the test project, the native LWJGL libraries were missing. With this commit the
native libraries are there, but it is only a workaround. The goal is to have all
native libraries inside this lib here, so everyone can use it, independent of
the OS. (It's a todo for the future)